Rename test_vsync_waiters to vsync_waiters_test for consistency.
