+++ shell/common/BUILD.gn
@@ -167,8 +167,8 @@ if (current_toolchain == host_toolchain) {
       "shell_test.cc",
       "shell_test.h",
       "shell_unittests.cc",
-      "test_vsync_waiters.cc",
-      "test_vsync_waiters.h",
+      "vsync_waiters_test.cc",
+      "vsync_waiters_test.h",
     ]
 
     deps = [
