+++ shell/common/shell_test.h
@@ -13,8 +13,8 @@
 #include "flutter/lib/ui/window/platform_message.h"
 #include "flutter/shell/common/run_configuration.h"
 #include "flutter/shell/common/shell.h"
-#include "flutter/shell/common/test_vsync_waiters.h"
 #include "flutter/shell/common/thread_host.h"
+#include "flutter/shell/common/vsync_waiters_test.h"
 #include "flutter/shell/gpu/gpu_surface_gl_delegate.h"
 #include "flutter/testing/test_dart_native_resolver.h"
 #include "flutter/testing/test_gl_surface.h"
